LayoutTests/imported/w3c/web-platform-tests/css/css-inline/inheritance-expected.txt


PASS Property alignment-baseline has initial value baseline
PASS Property alignment-baseline does not inherit
FAIL Property baseline-shift has initial value 0px assert_equals: expected "0px" but got "baseline"
PASS Property baseline-shift does not inherit
PASS Property dominant-baseline has initial value auto
FAIL Property dominant-baseline inherits assert_equals: expected "central" but got "auto"
FAIL Property initial-letters has initial value normal assert_true: initial-letters doesn't seem to be supported in the computed style expected true got false
FAIL Property initial-letters does not inherit assert_true: expected true got false
FAIL Property initial-letters-align has initial value alphabetic assert_true: initial-letters-align doesn't seem to be supported in the computed style expected true got false
FAIL Property initial-letters-align inherits assert_true: initial-letters-align doesn't seem to be supported in the computed style expected true got false
FAIL Property initial-letters-wrap has initial value none assert_true: initial-letters-wrap doesn't seem to be supported in the computed style expected true got false
FAIL Property initial-letters-wrap inherits assert_true: initial-letters-wrap doesn't seem to be supported in the computed style expected true got false
FAIL Property initial-sizing has initial value normal assert_true: initial-sizing doesn't seem to be supported in the computed style expected true got false
FAIL Property initial-sizing does not inherit assert_true: expected true got false
PASS Property line-height has initial value normal
PASS Property line-height inherits
PASS Property vertical-align has initial value baseline
PASS Property vertical-align does not inherit
